requirements.txt
pyaudio==0.2.12
pydub==0.25.1
python-dotenv==0.21.0
requests==2.28.1
